% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.confirm.1.R
\name{f.confirm.1}
\alias{f.confirm.1}
\title{Confirm Stage Analyses on a Single Dataset}
\usage{
f.confirm.1(learn.obj, ac.alpha = 0.04, sg.alpha = 0.01, verbose = FALSE)
}
\arguments{
\item{learn.obj}{an R object from the function output of \code{\link{f.learn.1}}}

\item{ac.alpha}{alpha level assigned to (two-stage) all-comer analysis}

\item{sg.alpha}{alpha level assigned to (two-stage) sensitive group analysis}

\item{verbose}{(logical) whether to print out results for each trial (default: FALSE)}
}
\value{
a list of three numericals showing all-comer-only (one-stage) analysis, two-stage all-comer analysis and
two-stage subgroup analysis significance (1) or non-significance (0)
}
\description{
This function analyzes a single dataset based on the results from the learn stage, and prepares quantities for
the subsequent learn/confirm analysis.
}
\author{
Gu Mi <neo.migu@gmail.com>
}
\seealso{
\code{\link{f.learn.1}} for learn stage analysis for a single dataset
\code{\link{f.learn.confirm}} for the main function of learn/confirm analysis (taking on a mega dataset)
}

