% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.learn.1.R
\name{f.learn.1}
\alias{f.learn.1}
\title{Learning Stage Analyses on a Single Dataset}
\usage{
f.learn.1(data.1, learn.allocation = 30, cutoff = 0.4, side = "LT")
}
\arguments{
\item{data.1}{an individual dataset consisting of all variables needed in an ASD}

\item{learn.allocation}{percentage of all subjects allocated to the learn stage: specify 30 means
30\% of subjects are allocated to the learn stage, while 70\% of subjects to the confirm stage}

\item{cutoff}{biomarker cutoff value for learn-stage dataset: a numerical value}

\item{side}{how to define sensitive subgroup using the cutoff value. Default is "LT" (less than),
indicating subjects in the subgroup all have biomarker value < cutoff. Alternatively, "GT" (greater than),
"LE" (less or equal to) or "GE" (greater or equal to) can be specified.}
}
\value{
a list of intermediate data results to be used for subsequent confirm stage analysis
}
\description{
This function analyzes a single dataset and prepares quantities for the subsequent confirm stage analysis.
}
\author{
Gu Mi <neo.migu@gmail.com>
}
\seealso{
\code{\link{f.confirm.1}} for confirm stage analysis for a single dataset
\code{\link{f.learn.confirm}} for the main function of learn/confirm analysis
}

