% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.learn.confirm.R
\name{f.learn.confirm}
\alias{f.learn.confirm}
\title{Main Function for Learn-Confirm Analysis}
\usage{
f.learn.confirm(data, learn.allocation = 30, cutoff = 0.4,
  ac.alpha = 0.04, sg.alpha = 0.01, verbose = FALSE)
}
\arguments{
\item{data}{pass a mega dataset (including multiple individual datasets)}

\item{learn.allocation}{percentage of all subjects allocated to the learn stage}

\item{ac.alpha}{alpha level assigned to (two-stage) all-comer analysis}

\item{sg.alpha}{alpha level assigned to (two-stage) sensitive group analysis}

\item{verbose}{(logical) whether to print out results for each trial (default: FALSE)}
}
\value{
a list of three vectors for indicating trial results, and other quantities for the integrated power analyses
performed by the function \code{\link{f.pwr.cal}}
}
\description{
This function takes a mega dataset and returns test significances for one-stage design, two-stage design
(all-comer and sensitive group). This is in preparation for function \code{\link{f.pwr.cal}} where multiple combinations
are considered.
}
\author{
Gu Mi <neo.migu@gmail.com>
}
\seealso{
\code{\link{f.learn.1}} for learn stage analysis for a single dataset
\code{\link{f.confirm.1}} for confirm stage analysis for a single dataset
}

