% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.pwr.cal.R
\name{f.pwr.cal}
\alias{f.pwr.cal}
\title{Power Calculations based on Different Parameter Combinations}
\usage{
f.pwr.cal(data, learn.allocation.v, cutoff, ac.alpha.v, sg.alpha.v)
}
\arguments{
\item{data}{the mega dataset consisting of many individual datasets of the same number of subjects and variables}

\item{learn.allocation.v}{a vector of the percentages for subjects allocated to the learn stage}

\item{cutoff}{biomarker cutoff value for learn-stage dataset: a numerical value}

\item{ac.alpha.v}{a vector of alpha levels assigned to all-comer analysis}

\item{sg.alpha.v}{a vector of alpha levels assigned to sensitive group analysis}
}
\value{
a list of objects containing key information for plotting by function \code{\link{plot_simASD}}
}
\description{
This function takes on a mega dataset, candidate learn/confirm allocations, biomarker cutoff value,
candidate alpha levels for all-comer analysis and sensitive group analysis. Please note that this function takes
time to run as different combinations are considered. Suggest to run this function separately and store the results
locally. Plotting (by function \code{\link{plot_simASD}} will directly utilize the result returned by this function.
}
\author{
Gu Mi <neo.migu@gmail.com>
}
\seealso{
\code{\link{f.learn.1}} for learn stage analysis for a single dataset
\code{\link{f.confirm.1}} for confirm stage analysis for a single dataset
\code{\link{f.learn.confirm}} for the main function of learn/confirm analysis (taking on a mega dataset)
\code{\link{plot_simASD}} for plotting the final results
}

