% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simASD.R
\name{plot_simASD}
\alias{plot_simASD}
\title{Plot Empirical Powers vs. Learn/Confirm Allocations under Different Combinations of ASD Parameters}
\usage{
plot_simASD(fpc.obj)
}
\arguments{
\item{fpc.obj}{an object from the returned result of function \code{\link{f.pwr.cal}}}
}
\value{
a ggplot2 object with empirical power vs. learn/confirm allocation plot, with one-stage design power as a
horizontal dashed line for comparison
}
\description{
This function takes an object from the return result of function \code{\link{f.pwr.cal}}, and makes
a ggplot2 object showing the empirical powers vs. learn/confirm allocations under different alpha-splitting scenarios.
The one-stage design power is also plotted as a horizontal dashed line for comparison.
}
\author{
Gu Mi <neo.migu@gmail.com>
}
\seealso{
\code{\link{f.learn.confirm}} for the main function of learn/confirm analysis (taking on a mega dataset)
\code{\link{f.pwr.cal}} for calculating powers based on a mega dataset
}

