% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.cal.R
\name{power.cal}
\alias{power.cal}
\title{Detailed power calculations for one-stage, two-stage, two-stage all-comers, and two-stage sensitive group}
\usage{
power.cal(f.learn.confirm.obj)
}
\arguments{
\item{f.learn.confirm.obj}{an object from the function output \code{\link{f.learn.confirm}}}
}
\value{
a print out on console for powers from different designs and stages
}
\description{
This function provides an overview summary of the powers from different designs and stages, by taking
an object returned by the function \code{\link{f.learn.confirm}} for a particular learn/confirm allocation, biomarker
cutoff and a particular alpha-splitting.
}
\author{
Gu Mi <neo.migu@gmail.com>
}

