% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pwr.cal}
\alias{pwr.cal}
\title{helper functions for getting powers, given a particular learn and confirm object}
\usage{
pwr.cal(f.learn.confirm.obj)
}
\arguments{
\item{f.learn.confirm.obj}{an object from the function output \code{\link{f.learn.confirm}}}
}
\author{
Gu Mi <neo.migu@gmail.com>
}
\keyword{internal}

