% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simASD.R
\docType{package}
\name{simASD-package}
\alias{simASD}
\alias{simASD-package}
\title{Enhancement of the Adaptive Signature Design (ASD) for Learning and Confirming in a Single Pivotal Trial}
\description{
An R package for performing a simple two-stage adaptive signature design (ASD) based on simulated datasets
}
\author{
Gu Mi <neo.migu@gmail.com>

Maintainer: Gu Mi <https://github.com/gu-mi>
}
\keyword{package}

